# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Твой API Справочник Организаций",
        "description": "API для работы с организациями, зданиями и видами деятельности",
        "contact": {
            "email": "твой@email.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Локальный сервер"
        }
    ],
    "paths": {
...
        "/api/organizations/in-rectangle": {
            "get": {
                "tags": [
                    "Организации"
                ],
                "summary": "Поиск организаций в прямоугольной области",
                "description": "Находит организации, расположенные в указанной прямоугольной области на карте (bounding box). Возвращает список организаций с телефонными номерами.",
                "operationId": "getOrganizationsInRectangle",
                "parameters": [
                    {
                        "name": "api_key",
                        "in": "query",
                        "description": "Ключ API для аутентификации.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "123"
                    },
                    {
                        "name": "sw_lat",
                        "in": "query",
                        "description": "Широта юго-западного угла области (от -90 до 90).",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "maximum": 90,
                            "minimum": -90
                        },
                        "example": 55.7
                    },
                    {
                        "name": "sw_lng",
                        "in": "query",
                        "description": "Долгота юго-западного угла области (от -180 до 180).",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "maximum": 180,
                            "minimum": -180
                        },
                        "example": 37.5
                    },
                    {
                        "name": "ne_lat",
                        "in": "query",
                        "description": "Широта северо-восточного угла области (от -90 до 90).",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "maximum": 90,
                            "minimum": -90
                        },
                        "example": 55.8
                    },
                    {
                        "name": "ne_lng",
                        "in": "query",
                        "description": "Долгота северо-восточного угла области (от -180 до 180).",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "maximum": 180,
                            "minimum": -180
                        },
                        "example": 37.6
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успех. Список организаций",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/OrganizationWithPhonesResource"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Не найдено",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Организации не найдены"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The sw lat field is required."
                                        },
                                        "errors": {
                                            "properties": {
                                                "sw_lat": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The sw lat field is required."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/organizations/activity-tree": {
            "get": {
                "tags": [
                    "Организации"
                ],
                "summary": "Поиск организаций по иерархии видов деятельности",
                "description": "Находит организации, связанные с видом деятельности и всеми его потомками. Возвращает организации с телефонами, зданиями и деревом видов деятельности.",
                "operationId": "getOrgsByActivityTree",
                "parameters": [
                    {
                        "name": "api_key",
                        "in": "query",
                        "description": "Ключ API для аутентификации.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "123"
                    },
                    {
                        "name": "activity_name",
                        "in": "query",
                        "description": "Название вида деятельности (минимум 3 символа). Например: 'вет' для ветеринарии.",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "maxLength": 255,
                            "minLength": 3
                        },
                        "example": "вет"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успех. Организации с деревом видов деятельности",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/FullOrganizationResource"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Не найдено",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Виды деятельности не найдены"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The activity name must be at least 3 characters."
                                        },
                                        "errors": {
                                            "properties": {
                                                "activity_name": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The activity name must be at least 3 characters."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "ActivityResource": {
                "title": "Ресурс вида деятельности",
                "description": "Структура данных, представляющая вид экономической деятельности с иерархической организацией",
                "required": [
                    "id",
                    "name",
                    "depth"
                ],
                "properties": {
                    "id": {
                        "title": "Идентификатор",
                        "description": "Уникальный числовой идентификатор вида деятельности",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "title": "Наименование",
                        "description": "Официальное название вида экономической деятельности",
                        "type": "string",
                        "example": "Производство промышленного оборудования"
                    },
                    "depth": {
                        "title": "Уровень вложенности",
                        "description": "Порядковый уровень в иерархической структуре видов деятельности (0 - корневой уровень)",
                        "type": "integer",
                        "example": 0
                    },
                    "children": {
                        "title": "Дочерние элементы",
                        "description": "Массив подчиненных видов деятельности в иерархической структуре (рекурсивная структура)",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ActivityResource"
                        }
                    }
                },
                "type": "object"
            },
...
            "BuildingResource": {
                "title": "Ресурс здания",
                "description": "Структура данных, представляющая информацию о здании",
                "required": [
                    "id",
                    "address",
                    "coordinates"
                ],
                "properties": {
                    "id": {
                        "title": "Идентификатор здания",
                        "description": "Уникальный числовой идентификатор объекта недвижимости",
                        "type": "integer",
                        "example": 1
                    },
                    "address": {
                        "title": "Адрес",
                        "description": "Физический адрес местонахождения здания",
                        "type": "string",
                        "example": "ул. Тверская, 15"
                    },
                    "coordinates": {
                        "title": "Географические координаты",
                        "description": "Геопозиция здания в системе координат WGS 84",
                        "required": [
                            "lat",
                            "lng"
                        ],
                        "properties": {
                            "lat": {
                                "title": "Широта",
                                "description": "Географическая широта в десятичных градусах",
                                "type": "number",
                                "format": "float",
                                "example": 55.7558
                            },
                            "lng": {
                                "title": "Долгота",
                                "description": "Географическая долгота в десятичных градусах",
                                "type": "number",
                                "format": "float",
                                "example": 37.6173
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "FullOrganizationResource": {
                "title": "Полный ресурс организации",
                "description": "Комплексное представление данных об организации",
                "properties": {
                    "organization": {
                        "title": "Данные организации",
                        "description": "Объект, содержащий полную информацию об организации",
                        "type": "object",
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/OrganizationResource"
                            },
                            {
                                "required": [
                                    "phones",
                                    "building",
                                    "activities"
                                ],
                                "properties": {
                                    "phones": {
                                        "title": "Контактные телефоны",
                                        "description": "Список контактных номеров телефонов организации",
                                        "type": "array",
                                        "items": {
                                            "$ref": "#/components/schemas/OrganizationPhoneResource"
                                        }
                                    },
                                    "building": {
                                        "$ref": "#/components/schemas/BuildingResource"
                                    },
                                    "activities": {
                                        "title": "Виды деятельности",
                                        "description": "Перечень видов экономической деятельности организации",
                                        "type": "array",
                                        "items": {
                                            "$ref": "#/components/schemas/ActivityResource"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        ]
                    }
                },
                "type": "object"
            },
            "OrganizationPhoneResource": {
                "title": "Ресурс телефонного номера организации",
                "description": "Представление контактного телефонного номера организации",
                "required": [
                    "id",
                    "number"
                ],
                "properties": {
                    "id": {
                        "title": "Идентификатор",
                        "description": "Уникальный числовой идентификатор телефонного номера",
                        "type": "integer",
                        "example": 1
                    },
                    "number": {
                        "title": "Номер телефона",
                        "description": "Контактный телефонный номер в международном или национальном формате",
                        "type": "string",
                        "example": "+7 (495) 123-45-67"
                    }
                },
                "type": "object"
            },
            "OrganizationResource": {
                "title": "Базовый ресурс организации",
                "description": "Минимальный набор данных для идентификации организации",
                "required": [
                    "id",
                    "name"
                ],
                "properties": {
                    "id": {
                        "title": "Идентификатор организации",
                        "description": "Уникальный числовой идентификатор организации в системе",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "title": "Наименование",
                        "description": "Официальное название организации",
                        "type": "string",
                        "example": "ООО 'Промышленные технологии'"
                    }
                },
                "type": "object"
            },
            "OrganizationWithPhonesResource": {
                "title": "Ресурс организации с телефонами",
                "description": "Данные организации, включая идентификатор, наименование и список телефонных номеров.",
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/OrganizationResource"
                    },
                    {
                        "properties": {
                            "phones": {
                                "title": "Телефоны",
                                "description": "Список телефонных номеров, связанных с организацией.",
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/OrganizationPhoneResource"
                                }
                            }
                        },
                        "type": "object"
                    }
                ]
            }
        },
        "securitySchemes": {
            "api_key": {
                "type": "apiKey",
                "name": "api_key",
                "in": "query"
            }
        }
    },
...
}